requirements.txt
openpyxl==3.0.9
pillow==8.3.2
pandas==1.3.3
pathlib==1.0.1
transliterate==1.10.2